Update config.yaml
Fix an error in Rabbit and default parameters settled to match the original GoPredSim defaults as possible.

code line 34 in RabbitMQ: port changed to 5672 (to fix protocol error as described here: https://github.com/pika/pika/issues/1247)
code line 61, sets the default reference as GOA2022
code line 64, sets 1 closest hit. 
code line 91, the distance_threshold cannot be "0", 2 is settled as a reasonable value.
code line 95, the distance_threshold cannot be "0", 2 is settled as a reasonable value.
code line 99, the distance_threshold cannot be "0", 2 is settled as a reasonable value.